QueryServer: Abstract over the query running parts of the query server in preperation for the new query server.
