+++ extensions/ql-vscode/src/vscode-tests/no-workspace/run-queries.test.ts
@@ -4,13 +4,15 @@ import * as fs from 'fs-extra';
 import * as sinon from 'sinon';
 import { Uri } from 'vscode';
 
-import { QueryEvaluationInfo } from '../../run-queries';
-import { Severity, compileQuery } from '../../pure/messages';
+import { Severity, compileQuery, registerDatabases, deregisterDatabases } from '../../pure/legacy-messages';
 import * as config from '../../config';
 import { tmpDir } from '../../helpers';
-import { QueryServerClient } from '../../queryserver-client';
+import { QueryServerClient } from '../../legacy-query-server/queryserver-client';
 import { CodeQLCliServer } from '../../cli';
 import { SELECT_QUERY_NAME } from '../../contextual/locationFinder';
+import { QueryInProgress } from '../../legacy-query-server/run-queries';
+import { LegacyQueryRunner } from '../../legacy-query-server/legacyRunner';
+import { DatabaseItem } from '../../databases';
 
 describe('run-queries', () => {
   let sandbox: sinon.SinonSandbox;
@@ -29,55 +31,53 @@ describe('run-queries', () => {
     const info = createMockQueryInfo(true, saveDir);
 
     expect(info.compiledQueryPath).to.eq(path.join(saveDir, 'compiledQuery.qlo'));
-    expect(info.dilPath).to.eq(path.join(saveDir, 'results.dil'));
-    expect(info.resultsPaths.resultsPath).to.eq(path.join(saveDir, 'results.bqrs'));
-    expect(info.resultsPaths.interpretedResultsPath).to.eq(path.join(saveDir, 'interpretedResults.sarif'));
+    expect(info.queryEvalInfo.dilPath).to.eq(path.join(saveDir, 'results.dil'));
+    expect(info.queryEvalInfo.resultsPaths.resultsPath).to.eq(path.join(saveDir, 'results.bqrs'));
+    expect(info.queryEvalInfo.resultsPaths.interpretedResultsPath).to.eq(path.join(saveDir, 'interpretedResults.sarif'));
     expect(info.dbItemPath).to.eq(Uri.file('/abc').fsPath);
   });
 
   it('should check if interpreted results can be created', async () => {
     const info = createMockQueryInfo(true);
 
-    expect(info.canHaveInterpretedResults()).to.eq(true);
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '1').to.eq(true);
 
-    (info as any).databaseHasMetadataFile = false;
-    expect(info.canHaveInterpretedResults()).to.eq(false);
+    (info.queryEvalInfo as any).databaseHasMetadataFile = false;
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '2').to.eq(false);
 
-    (info as any).databaseHasMetadataFile = true;
+    (info.queryEvalInfo as any).databaseHasMetadataFile = true;
     info.metadata!.kind = undefined;
-    expect(info.canHaveInterpretedResults()).to.eq(false);
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '3').to.eq(false);
 
     info.metadata!.kind = 'table';
-    expect(info.canHaveInterpretedResults()).to.eq(false);
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '4').to.eq(false);
 
     // Graphs are not interpreted unless canary is set
     info.metadata!.kind = 'graph';
-    expect(info.canHaveInterpretedResults()).to.eq(false);
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '5').to.eq(false);
 
     (config.isCanary as sinon.SinonStub).returns(true);
-    expect(info.canHaveInterpretedResults()).to.eq(true);
+    expect(info.queryEvalInfo.canHaveInterpretedResults(), '6').to.eq(true);
   });
 
   [SELECT_QUERY_NAME, 'other'].forEach(resultSetName => {
     it(`should export csv results for result set ${resultSetName}`, async () => {
       const csvLocation = path.join(tmpDir.name, 'test.csv');
-      const qs = createMockQueryServerClient(
-        createMockCliServer({
-          bqrsInfo: [{ 'result-sets': [{ name: resultSetName }, { name: 'hucairz' }] }],
-          bqrsDecode: [{
-            columns: [{ kind: 'NotString' }, { kind: 'String' }],
-            tuples: [['a', 'b'], ['c', 'd']],
-            next: 1
-          }, {
-            // just for fun, give a different set of columns here
-            // this won't happen with the real CLI, but it's a good test
-            columns: [{ kind: 'String' }, { kind: 'NotString' }, { kind: 'StillNotString' }],
-            tuples: [['a', 'b', 'c']]
-          }]
-        })
-      );
+      const cliServer = createMockCliServer({
+        bqrsInfo: [{ 'result-sets': [{ name: resultSetName }, { name: 'hucairz' }] }],
+        bqrsDecode: [{
+          columns: [{ kind: 'NotString' }, { kind: 'String' }],
+          tuples: [['a', 'b'], ['c', 'd']],
+          next: 1
+        }, {
+          // just for fun, give a different set of columns here
+          // this won't happen with the real CLI, but it's a good test
+          columns: [{ kind: 'String' }, { kind: 'NotString' }, { kind: 'StillNotString' }],
+          tuples: [['a', 'b', 'c']]
+        }]
+      });
       const info = createMockQueryInfo();
-      const promise = info.exportCsvResults(qs, csvLocation);
+      const promise = info.queryEvalInfo.exportCsvResults(cliServer, csvLocation);
 
       const result = await promise;
       expect(result).to.eq(true);
@@ -86,14 +86,14 @@ describe('run-queries', () => {
       expect(csv).to.eq('a,"b"\nc,"d"\n"a",b,c\n');
 
       // now verify that we are using the expected result set
-      expect((qs.cliServer.bqrsDecode as sinon.SinonStub).callCount).to.eq(2);
-      expect((qs.cliServer.bqrsDecode as sinon.SinonStub).getCall(0).args[1]).to.eq(resultSetName);
+      expect((cliServer.bqrsDecode as sinon.SinonStub).callCount).to.eq(2);
+      expect((cliServer.bqrsDecode as sinon.SinonStub).getCall(0).args[1]).to.eq(resultSetName);
     });
   });
 
   it('should export csv results with characters that need to be escaped', async () => {
     const csvLocation = path.join(tmpDir.name, 'test.csv');
-    const qs = createMockQueryServerClient(
+    const cliServer =
       createMockCliServer({
         bqrsInfo: [{ 'result-sets': [{ name: SELECT_QUERY_NAME }, { name: 'hucairz' }] }],
         bqrsDecode: [{
@@ -109,10 +109,9 @@ describe('run-queries', () => {
             [123.98, 456.99],
           ],
         }]
-      })
-    );
+      });
     const info = createMockQueryInfo();
-    const promise = info.exportCsvResults(qs, csvLocation);
+    const promise = info.queryEvalInfo.exportCsvResults(cliServer, csvLocation);
 
     const result = await promise;
     expect(result).to.eq(true);
@@ -121,19 +120,18 @@ describe('run-queries', () => {
     expect(csv).to.eq('"a","""b"""\nc,xxx,"d,yyy"\naaa " bbb,"ccc "" ddd"\ntrue,"false"\n123,"456"\n123.98,"456.99"\n');
 
     // now verify that we are using the expected result set
-    expect((qs.cliServer.bqrsDecode as sinon.SinonStub).callCount).to.eq(1);
-    expect((qs.cliServer.bqrsDecode as sinon.SinonStub).getCall(0).args[1]).to.eq(SELECT_QUERY_NAME);
+    expect((cliServer.bqrsDecode as sinon.SinonStub).callCount).to.eq(1);
+    expect((cliServer.bqrsDecode as sinon.SinonStub).getCall(0).args[1]).to.eq(SELECT_QUERY_NAME);
   });
 
   it('should handle csv exports for a query with no result sets', async () => {
     const csvLocation = path.join(tmpDir.name, 'test.csv');
-    const qs = createMockQueryServerClient(
+    const cliServer =
       createMockCliServer({
         bqrsInfo: [{ 'result-sets': [] }]
-      })
-    );
+      });
     const info = createMockQueryInfo();
-    const result = await info.exportCsvResults(qs, csvLocation);
+    const result = await info.queryEvalInfo.exportCsvResults(cliServer, csvLocation);
     expect(result).to.eq(false);
   });
 
@@ -187,9 +185,126 @@ describe('run-queries', () => {
     });
   });
 
+
+  describe('register', () => {
+    it('should register', async () => {
+      const qs = createMockQueryServerClient(
+        {
+          cliConstraints: {
+            supportsDatabaseRegistration: () => true
+          }
+        } as any);
+      const runner = new LegacyQueryRunner(qs);
+      const mockProgress = 'progress-monitor';
+      const mockCancel = 'cancel-token';
+      const datasetUri = Uri.file('dataset-uri');
+
+      const dbItem: DatabaseItem = {
+        contents: {
+          datasetUri
+        }
+      } as any;
+
+      await runner.registerDatabase(mockProgress as any, mockCancel as any, dbItem);
+
+      expect(qs.sendRequest).to.have.been.calledOnceWith(
+        registerDatabases,
+        {
+          databases: [
+            {
+              dbDir: datasetUri.fsPath,
+              workingSet: 'default'
+            }
+          ]
+        },
+        mockCancel,
+        mockProgress
+      );
+    });
+
+    it('should deregister', async () => {
+      const qs = createMockQueryServerClient(
+        {
+          cliConstraints: {
+            supportsDatabaseRegistration: () => true
+          }
+        } as any);
+      const runner = new LegacyQueryRunner(qs);
+      const mockProgress = 'progress-monitor';
+      const mockCancel = 'cancel-token';
+      const datasetUri = Uri.file('dataset-uri');
+
+      const dbItem: DatabaseItem = {
+        contents: {
+          datasetUri
+        }
+      } as any;
+
+      await runner.deregisterDatabase(mockProgress as any, mockCancel as any, dbItem);
+
+      expect(qs.sendRequest).to.have.been.calledOnceWith(
+        deregisterDatabases,
+        {
+          databases: [
+            {
+              dbDir: datasetUri.fsPath,
+              workingSet: 'default'
+            }
+          ]
+        },
+        mockCancel,
+        mockProgress
+      );
+    });
+
+    it('should not register if unsupported', async () => {
+      const qs = createMockQueryServerClient(
+        {
+          cliConstraints: {
+            supportsDatabaseRegistration: () => false
+          }
+        } as any);
+      const runner = new LegacyQueryRunner(qs);
+      const mockProgress = 'progress-monitor';
+      const mockCancel = 'cancel-token';
+      const datasetUri = Uri.file('dataset-uri');
+
+      const dbItem: DatabaseItem = {
+        contents: {
+          datasetUri
+        }
+      } as any;
+      await runner.registerDatabase(mockProgress as any, mockCancel as any, dbItem);
+      expect(qs.sendRequest).not.to.have.been.called;
+    });
+
+    it('should not deregister if unsupported', async () => {
+      const qs = createMockQueryServerClient(
+        {
+          cliConstraints: {
+            supportsDatabaseRegistration: () => false
+          }
+        } as any);
+      const runner = new LegacyQueryRunner(qs);
+      const mockProgress = 'progress-monitor';
+      const mockCancel = 'cancel-token';
+      const datasetUri = Uri.file('dataset-uri');
+
+      const dbItem: DatabaseItem = {
+        contents: {
+          datasetUri
+        }
+      } as any;
+      await runner.registerDatabase(mockProgress as any, mockCancel as any, dbItem);
+      expect(qs.sendRequest).not.to.have.been.called;
+    });
+
+  });
+
+
   let queryNum = 0;
   function createMockQueryInfo(databaseHasMetadataFile = true, saveDir = `save-dir${queryNum++}`) {
-    return new QueryEvaluationInfo(
+    return new QueryInProgress(
       saveDir,
       Uri.parse('file:///abc').fsPath,
       databaseHasMetadataFile,
